requirements.txt
binance-connector==3.5.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
frozendict==2.3.10
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
moralis==0.1.42
numpy==1.26.2
pandas==2.1.4
pycryptodome==3.19.0
python-dateutil==2.8.2
python-environ==0.4.54
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
typing_extensions==4.3.0
tzdata==2023.3
urllib3==1.26.18
websocket-client==1.7.0
Werkzeug==3.0.1
